requirements.txt
torch
torchvision
transformers
datasets
opencv-python
facenet-pytorch        # MTCNN face detection
scikit-learn
streamlit
Pillow
tqdm
pandas
numpy
matplotlib
seaborn
accelerate            # For efficient training